move logic to internal folder
